requirements.txt
django
gunicorn
django-heroku
darkskylib
pytz